Implemented the Universal Chess Interface (UCI) Protocol
Description: The UCI protocol is a base set of rules and standards that chess GUI’s use to communicate with chess engines. Chess engines must be able to accept requests made from the GUI that follow the UCI protocol. This commit sets up the project for connecting to a GUI later.

Explanation of uci.c
The UCI_Loop listens for requests made from the gui/user. ParsePosition is executed when moves are made and the board state is updated and ParseGo searches the next best move for the engine to take based on input parameters (white time, black time, movestogo, search depth, time left, etc.)

Explanation misc.c
We have to include the ReadInput() and InputWaiting() functions to peek and see if there is an input waiting from the gui. If there is, then we must stop the engine, or if we see that there is a quit message, we must stop the engine. We have to peek the engine input otherwise the engine will keep on running.