Added Chess Engine Communication Protocol(xboard/winboard); Can play with both a gui and in the console
XBoard/WinBoard is a method/protocol for users to send messages to engine through console. Makes it so that engine is aware of the game situation. The engine has more control over the game. Allows user to play against Claude through the console. When you play the engine, you have to specify the depth for the engine to search and the time the engine gets. You can type help to get a list of commands and how to play.