Transposition Table (Hash Table); Further improvements to the search function
Database that stores the results of previously performed searches. Greatly reduces the search space of a chess tree. Make it so that you don’t evaluate the same moves over and over again (transposition). We can ignore some branches. We do this by changing the pvtable structure into a hashtable. The hashtable allows you to recall the past moves and the information from the past moves quickly, allowing the engine to make better decisions.